Some more platform independence
* is not compiled yet
* use intptr_t from stdint.h as int <-> (void*)
* some minor warnings fixed
* fix 32Bit hard dependency in BinFile.hc.c
  (dunno if that matters, actually, but it was wrong)

+++ src/lib/BasicTypes/Char.hc.h
@@ -9,7 +9,7 @@
 
 typedef unsigned char OCSCHAR;
 
-#define pack_char(x)   ((OBJ)((WORD)(((OCSCHAR)(x)<<1)+1)))
+#define pack_char(x)   ((OBJ)((intptr_t)(((OCSCHAR)(x)<<1)+1)))
 #define unpack_char(x) ((OCSCHAR)((WORD)(x)>>1))
 #define convert2char(x) ((WORD)(x))
